llm/session_test.go
package llm

import (
	"reflect"
	"testing"
)

func TestCleanupExpiredSessions(t *testing.T) {
	tests := []struct {
		name string
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			CleanupExpiredSessions()
		})
	}
}

func TestUpdateSessionHistory(t *testing.T) {
	type args struct {
		sessionID string
		role      string
		content   string
	}
	tests := []struct {
		name string
		args args
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			UpdateSessionHistory(tt.args.sessionID, tt.args.role, tt.args.content)
		})
	}
}

func TestGetSessionHistory(t *testing.T) {
	type args struct {
		sessionID string
	}
	tests := []struct {
		name string
		args args
		want []Message
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := GetSessionHistory(tt.args.sessionID); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("GetSessionHistory() = %v, want %v", got, tt.want)
			}
		})
	}
}
